"removing files with secrets"ttprequest troubles

+++ UVM1.5/Controllers/ItemController.cs
@@ -370,13 +370,8 @@ namespace UVM1._5.Controllers
 
             }
 
-            /*OpenAIController ai = new OpenAIController();
-            string itemDesc = $"This photos shows a {item.Brand.Name} {item.Model} {item.Category.Name} ";
-
-            foreach(var img in item.Images)
-            {
-                ai.MatchImageToDescriptionAsync(img, itemDesc);
-            }*/
+            OpenAIController ai = new OpenAIController();
+            ai.GPTVision();
 
             TempFile(item.Images[0], "");
 
@@ -399,21 +394,26 @@ namespace UVM1._5.Controllers
 
         public void TempFile(byte[] img, string path)
         {
-            byte[] filedata = img;
-            string extension = "jpg"; // "pdf", etc
+            if(img != null)
+            {
+                byte[] filedata = img;
+                string extension = "jpg"; // "pdf", etc
 
 
-            //string filename = System.IO.Path.GetTempFileName() + "." + extension; // Makes something like "C:\Temp\blah.tmp.pdf"
-            string filename = "wwwroot/images/temp.jpg";
-            System.IO.File.WriteAllBytes(filename,filedata);
+                //string filename = System.IO.Path.GetTempFileName() + "." + extension; // Makes something like "C:\Temp\blah.tmp.pdf"
+                string filename = "wwwroot/images/temp.jpg";
+                System.IO.File.WriteAllBytes(filename, filedata);
 
-            /*var process = Process.Start(filename);
-            // Clean up our temporary file...
-            process.Exited += (s, e) => System.IO.File.Delete(filename);*/
-            System.Diagnostics.Debug.WriteLine(filename);
-            //System.IO.File.Delete(filename);
-            System.Diagnostics.Debug.WriteLine("\n\n\nplease work");
-            //System.IO.File.Delete(filename);
+                /*var process = Process.Start(filename);
+                // Clean up our temporary file...
+                process.Exited += (s, e) => System.IO.File.Delete(filename);*/
+                System.Diagnostics.Debug.WriteLine(filename);
+                //System.IO.File.Delete(filename);
+                System.Diagnostics.Debug.WriteLine("\n\n\nplease work");
+                //System.IO.File.Delete(filename);
+
+            }
+
         }
     }
 }
